backend/requirements.txt
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
Deprecated==1.2.14
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.6.0
flask-marshmallow==1.2.1
flask-talisman==1.1.0
flask-validators==1.0
greenlet==3.0.3
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
langid==1.1.6
limits==3.11.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
numpy==1.26.4
ordered-set==4.1.0
packaging==24.0
psycopg==3.1.18
psycopg-binary==3.1.18
Pygments==2.17.2
PyJWT==2.8.0
python-dotenv==1.0.1
rich==13.7.1
typing_extensions==4.11.0
Werkzeug==3.0.2
wrapt==1.16.0
